Update CharacterBase.cs
Changed health variables from floats to ints for future use with stats and status effects

+++ Assets/Scripts/Entity/CharacterBase.cs
@@ -18,8 +18,8 @@ public class CharacterBase : MonoBehaviour
     public int quick;
     public int strong;
     public int toughness;
-    public float maxHealth;
-    protected float currentHealth;
+    public int maxHealth;
+    protected int currentHealth;
     [Tooltip(" Defense = [Quick –Armor’s Impeding value]")]
     public int defence;
     [Tooltip(" Pain Threshold = Strong/2 (rounded up)")]
